+++ dart/lib/libwallet.dart
@@ -82,6 +82,8 @@ export 'src/models/signed_message.dart' show SignedMessage;
 export 'src/models/remote_key_session.dart'
     show RemoteKeySession, RemoteKeyValidation;
 export 'src/models/wallet_backup.dart' show WalletBackupEntry;
+export 'src/models/device_transfer.dart'
+    show DeviceTransferSession, DeviceShareEntry, DeviceTransferImportResult;
 export 'src/models/nft_listing.dart' show NftListing;
 export 'src/models/rpc_test.dart' show RpcTestResult;
 export 'src/models/transaction_simulation.dart'